requirements.txt
pandas
numpy
dvc
keras
scikit-learn
matplotlib
seaborn
tensorflow
joblib
PyYaml
streamlit
mlflow